requirements.txt
fastmcp==2.10.6
matplotlib==3.10.3
numpy==2.3.1
Pillow==11.3.0
Requests==2.32.4
websocket_client==1.8.0
